+++ src/xrGame/ai/stalker/ai_stalker.cpp
@@ -114,7 +114,7 @@ void CAI_Stalker::reinit			()
 	sound().sound_prefix			(SpecificCharacter().sound_voice_prefix());
 
 #ifdef DEBUG_MEMORY_MANAGER
-	u32								start = 0;
+	size_t								start = 0;
 	if (g_bMEMO)
 		start						= Memory.mem_usage();
 #endif // DEBUG_MEMORY_MANAGER
@@ -123,7 +123,7 @@ void CAI_Stalker::reinit			()
 
 #ifdef DEBUG_MEMORY_MANAGER
 	if (g_bMEMO)
-		Msg					("CAI_Stalker::LoadSounds() : %d",Memory.mem_usage() - start);
+		Msg					("CAI_Stalker::LoadSounds() : %lld",Memory.mem_usage() - start);
 #endif // DEBUG_MEMORY_MANAGER
 
 	m_pPhysics_support->in_Init		();
@@ -238,7 +238,7 @@ void CAI_Stalker::reload			(LPCSTR section)
 
 #ifdef DEBUG_MEMORY_MANAGER
 	if (g_bMEMO)
-		Msg					("brain().setup() : %d",Memory.mem_usage() - start);
+		Msg					("brain().setup() : %lld",Memory.mem_usage() - start);
 #endif // DEBUG_MEMORY_MANAGER
 
 	CCustomMonster::reload			(section);
@@ -529,7 +529,7 @@ void CAI_Stalker::Load				(LPCSTR section)
 BOOL CAI_Stalker::net_Spawn			(CSE_Abstract* DC)
 {
 #ifdef DEBUG_MEMORY_MANAGER
-	u32								start = 0;
+	size_t								start = 0;
 	if (g_bMEMO)
 		start						= Memory.mem_usage();
 #endif // DEBUG_MEMORY_MANAGER
@@ -553,7 +553,7 @@ BOOL CAI_Stalker::net_Spawn			(CSE_Abstract* DC)
 	set_money						(tpHuman->m_dwMoney, false);
 
 #ifdef DEBUG_MEMORY_MANAGER
-	u32									_start = 0;
+	size_t									_start = 0;
 	if (g_bMEMO)
 		_start							= Memory.mem_usage();
 #endif // DEBUG_MEMORY_MANAGER
@@ -562,7 +562,7 @@ BOOL CAI_Stalker::net_Spawn			(CSE_Abstract* DC)
 
 #ifdef DEBUG_MEMORY_MANAGER
 	if (g_bMEMO)
-		Msg					("CStalkerAnimationManager::reload() : %d",Memory.mem_usage() - _start);
+		Msg					("CStalkerAnimationManager::reload() : %lld",Memory.mem_usage() - _start);
 #endif // DEBUG_MEMORY_MANAGER
 
 	movement().m_head.current.yaw	= movement().m_head.target.yaw = movement().m_body.current.yaw = movement().m_body.target.yaw	= angle_normalize_signed(-tpHuman->o_torso.yaw);
@@ -645,7 +645,7 @@ BOOL CAI_Stalker::net_Spawn			(CSE_Abstract* DC)
 
 #ifdef DEBUG_MEMORY_MANAGER
 	if (g_bMEMO) {
-		Msg							("CAI_Stalker::net_Spawn() : %d",Memory.mem_usage() - start);
+		Msg							("CAI_Stalker::net_Spawn() : %lld",Memory.mem_usage() - start);
 	}
 #endif // DEBUG_MEMORY_MANAGER
 
@@ -1196,7 +1196,7 @@ CMemoryManager *CAI_Stalker::create_memory_manager		()
 DLL_Pure *CAI_Stalker::_construct			()
 {
 #ifdef DEBUG_MEMORY_MANAGER
-	u32									start = 0;
+	size_t									start = 0;
 	if (g_bMEMO)
 		start							= Memory.mem_usage();
 #endif // DEBUG_MEMORY_MANAGER
@@ -1215,7 +1215,7 @@ DLL_Pure *CAI_Stalker::_construct			()
 
 #ifdef DEBUG_MEMORY_MANAGER
 	if (g_bMEMO)
-		Msg								("CAI_Stalker::_construct() : %d",Memory.mem_usage() - start);
+		Msg								("CAI_Stalker::_construct() : %lld",Memory.mem_usage() - start);
 #endif // DEBUG_MEMORY_MANAGER
 
 	return								(this);
